refactor: change only support redis and redis cluster
